+++ test/recording_studio_support_test.rb
@@ -145,12 +145,13 @@ class RecordingStudioSupportTest < Minitest::Test
     default_layout_head = File.read(head_path)
 
     assert_includes default_layout_head, 'stylesheet_link_tag "flat_pack/application"'
-    assert_includes default_layout_head, 'start_with?("RecordingStudioSupport::", "RecordingStudioAdmin::")'
+    assert_includes default_layout_head, "start_with?"
     assert_includes default_layout_head, "recording_studio_root_switch_dropdown"
     assert_includes default_layout_head, "recording_studio_page_nav_right"
     assert_includes default_layout_head, "user_signed_in?"
     assert_includes default_layout_head, "RecordingStudioSupport::"
     assert_includes default_layout_head, "RecordingStudioAdmin::"
+    assert_includes default_layout_head, "RecordingStudioMoveable::"
     assert_includes default_layout_head, "Sign out"
     assert_includes default_layout_head, "destroy_user_session_path"
     assert_includes default_layout_head, "turbo_method: :delete"